+++ CHANGELOG.md
@@ -90,6 +90,8 @@ and this project adheres to
   - [#1568](https://github.com/iovisor/bpftrace/issues/1568)
 - Fix LNOT
   - [#1570](https://github.com/iovisor/bpftrace/pull/1570)
+- Fix invalid cast handling in tuple
+  - [#1572](https://github.com/iovisor/bpftrace/pull/1572)
 
 #### Tools
 - Hook up execsnoop.bt script onto `execveat` call